requirements.txt
python-telegram-bot==20.0  # Для работы с Telegram API
requests==2.28.1             # Для выполнения HTTP-запросов
python-dotenv==1.0.0         # Для загрузки переменных окружения из .env файла
